fix: adjust pooler import
